+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/V1beta1CertificatesAPIGroupClient.java
@@ -18,7 +18,7 @@ package io.fabric8.kubernetes.client;
 import io.fabric8.kubernetes.api.model.certificates.v1beta1.CertificateSigningRequest;
 import io.fabric8.kubernetes.api.model.certificates.v1beta1.CertificateSigningRequestList;
 import io.fabric8.kubernetes.client.dsl.NonNamespaceOperation;
-import io.fabric8.kubernetes.client.dsl.Resource;
+import io.fabric8.kubernetes.client.dsl.V1beta1CertificateSigningRequestResource;
 import io.fabric8.kubernetes.client.dsl.V1beta1CertificatesAPIGroupDSL;
 import io.fabric8.kubernetes.client.extension.ClientAdapter;
 
@@ -26,8 +26,10 @@ public class V1beta1CertificatesAPIGroupClient extends ClientAdapter<V1beta1Cert
     implements V1beta1CertificatesAPIGroupDSL {
 
   @Override
-  public NonNamespaceOperation<CertificateSigningRequest, CertificateSigningRequestList, Resource<CertificateSigningRequest>> certificateSigningRequests() {
-    return resources(CertificateSigningRequest.class, CertificateSigningRequestList.class);
+  public NonNamespaceOperation<CertificateSigningRequest, CertificateSigningRequestList, V1beta1CertificateSigningRequestResource<CertificateSigningRequest>> certificateSigningRequests() {
+    return (NonNamespaceOperation<CertificateSigningRequest, CertificateSigningRequestList, V1beta1CertificateSigningRequestResource<CertificateSigningRequest>>) resources(
+        CertificateSigningRequest.class, CertificateSigningRequestList.class,
+        V1beta1CertificateSigningRequestResource.class);
   }
 
   @Override